Separation components by concern
각각의 컴포넌트들을 역할에 따라 분리했습니다.
