+++ src/Title.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Title() {
+  return <div>Counter</div>;
+}
+
+export default Title;
